feat: rotate class
